requirements.txt
Django==1.8.5
django-cors-headers==1.1.0
django-resized==0.3.5
django-rest-multiple-models==1.6.1
django-stdimage==2.3.2
djangorestframework==3.3.2
Pillow==3.1.1
progressbar2==3.6.2
psycopg2==2.6.1
Pygments==2.1
wheel==0.24.0
